+++ ChangeLog.md
@@ -43,6 +43,7 @@ Other enhancements:
   `--work-dir` option is not a valid relative path.
 * Stack will use the value of the `GH_TOKEN`, or `GITHUB_TOKEN`, environment
   variable as credentials to authenticate its GitHub REST API requests.
+* `stack uninstall` also shows how to uninstall Stack-supplied tools.
 
 Bug fixes: